+++ deployment/terraform/batch.tf
@@ -1,7 +1,7 @@
 resource "aws_batch_compute_environment" "default" {
   depends_on = ["aws_iam_role_policy_attachment.batch_policy"]
 
-  compute_environment_name = "batch${var.environment}DefaultComputeEnvironment"
+  compute_environment_name = "batch${local.short}DefaultComputeEnvironment"
   type                     = "MANAGED"
   state                    = "ENABLED"
   service_role             = "${aws_iam_role.container_instance_batch.arn}"
@@ -41,7 +41,7 @@ resource "aws_batch_compute_environment" "default" {
 }
 
 resource "aws_batch_job_queue" "default" {
-  name                 = "queue${var.environment}Default"
+  name                 = "queue${local.short}Default"
   priority             = 1
   state                = "ENABLED"
   compute_environments = ["${aws_batch_compute_environment.default.arn}"]
@@ -70,11 +70,14 @@ data "template_file" "default_job_definition" {
     rollbar_server_side_access_token = "${var.rollbar_server_side_access_token}"
 
     aws_region = "${var.aws_region}"
+
+    batch_job_queue_name = "queue${local.short}Default"
+    batch_job_def_name = "job${local.short}Default"
   }
 }
 
 resource "aws_batch_job_definition" "default" {
-  name = "job${var.environment}Default"
+  name = "job${local.short}Default"
   type = "container"
 
   container_properties = "${data.template_file.default_job_definition.rendered}"