+++ deployment/terraform/variables.tf
@@ -1,5 +1,13 @@
+locals {
+  short = "${replace(var.project, " ", "")}${var.environment}"
+}
+
 variable "project" {
-  default = "Open Apparel Registry"
+  default = "Open Supply Hub"
+}
+
+variable "short_project" {
+  default = "osh"
 }
 
 variable "environment" {
@@ -85,7 +93,7 @@ variable "rds_auto_minor_version_upgrade" {
 }
 
 variable "rds_final_snapshot_identifier" {
-  default = "oar-rds-snapshot"
+  default = "osh-rds-snapshot"
 }
 
 variable "rds_monitoring_interval" {