+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Show sectors on facility detail sidebar [#1898](https://github.com/open-apparel-registry/open-apparel-registry/pull/1898)
 - Add sector search controls [#1899](https://github.com/open-apparel-registry/open-apparel-registry/pull/1899)
 - Display claim information [#1905](https://github.com/open-apparel-registry/open-apparel-registry/pull/1905)
+- Create OGR staging environment [#1885](https://github.com/open-apparel-registry/open-apparel-registry/pull/1885)
 
 ### Changed